updated get command
