+++ backend/cmd/users_export.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/gtsteffaniak/filebrowser/storage"
+	"github.com/gtsteffaniak/filebrowser/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -15,11 +16,11 @@ var usersExportCmd = &cobra.Command{
 	Long: `Export all users to a json or yaml file. Please indicate the
 path to the file where you want to write the users.`,
 	Args: jsonYamlArg,
-	Run: python(func(cmd *cobra.Command, args []string, store *storage.Storage) {
+	Run: initDb(func(cmd *cobra.Command, args []string, store *storage.Storage) {
 		list, err := d.store.Users.Gets("")
-		checkErr("d.store.Users.Gets", err)
+		utils.CheckErr("d.store.Users.Gets", err)
 
 		err = marshal(args[0], list)
-		checkErr("marshal", err)
+		utils.CheckErr("marshal", err)
 	}, pythonConfig{}),
 }
